requirements.txt
langgraph
langchain
langchain-google-genai
python-telegram-bot
google-api-python-client
google-auth-oauthlib
python-dotenv
requests
fastapi
uvicorn[standard]
langgraph-checkpoint-sqlite 